# Global warped motion compensation