--- a/Source/Lib/Common/Codec/EbAdaptiveMotionVectorPrediction.h
+++ b/Source/Lib/Common/Codec/EbAdaptiveMotionVectorPrediction.h
@@ -305,6 +305,13 @@ extern "C" {
         BlockSize bsize, int mib_size_log2);
     int is_inside_tile_boundary(TileInfo *tile, int16_t mvx, int16_t mvy, int mi_col, int mi_row, BlockSize bsize);
 
+    IntMv gm_get_motion_vector_enc(
+        const EbWarpedMotionParams *gm,
+        int32_t allow_hp,
+        BlockSize bsize,
+        int32_t mi_col, int32_t mi_row,
+        int32_t is_integer);
+
 #ifdef __cplusplus
 }
 #endif