--- a/Source/Lib/Common/Codec/EbDefinitions.h
+++ b/Source/Lib/Common/Codec/EbDefinitions.h
@@ -72,6 +72,7 @@ extern "C" {
 #define PRED_CHANGE_5L               1 // Change the MRP in 5L Pictures 3, 5 , 7 and 9 use 1 as the reference, 11, 13, 15 and 17 use 9 as the reference
 #define PRED_CHANGE_MOD              1 // Reorder the references for MRP
 #define SPEED_OPT                    1 // Speed optimization(s)
+#define GLOBAL_WARPED_MOTION         1 // Global warped motion detection and insertion
 
 #ifndef NON_AVX512_SUPPORT
 #define NON_AVX512_SUPPORT
@@ -196,15 +197,15 @@ enum {
 #if II_COMP_FLAG
 #if OBMC_FLAG
 #if FILTER_INTRA_FLAG
-#define MAX_NFL                                 110 // Maximum number of candidates MD can support
+#define MAX_NFL                                 125 // Maximum number of candidates MD can support
 #else
-#define MAX_NFL                                 105 // Maximum number of candidates MD can support
+#define MAX_NFL                                 120 // Maximum number of candidates MD can support
 #endif
 #else
-#define MAX_NFL                                  80
+#define MAX_NFL                                  95
 #endif
 #else
-#define MAX_NFL                                   65
+#define MAX_NFL                                   80
 #endif
 #define MAX_NFL_BUFF                              (MAX_NFL + CAND_CLASS_TOTAL)  //need one extra temp buffer for each fast loop call
 #define MAX_LAD                                   120 // max lookahead-distance 2x60fps
@@ -543,6 +544,7 @@ typedef enum CAND_CLASS {
 #if PAL_CLASS
     CAND_CLASS_7,
 #endif
+    CAND_CLASS_8,
     CAND_CLASS_TOTAL
 } CAND_CLASS;
 
@@ -1933,6 +1935,9 @@ typedef struct LoopFilterInfoN
 #define GM_TRANS_MIN -GM_TRANS_MAX
 #define GM_ALPHA_MIN -GM_ALPHA_MAX
 #define GM_ROW3HOMO_MIN -GM_ROW3HOMO_MAX
+
+#define USE_CUR_GM_REFMV 1
+
 /* clang-format off */
 typedef enum TransformationType
 {