--- a/Source/Lib/Common/Codec/EbGlobalMotionEstimationCost.h
+++ b/Source/Lib/Common/Codec/EbGlobalMotionEstimationCost.h
@@ -1,0 +1,26 @@
+/*
+* Copyright(c) 2019 Intel Corporation
+* SPDX - License - Identifier: BSD - 2 - Clause - Patent
+*/
+
+/*
+* Copyright (c) 2016, Alliance for Open Media. All rights reserved
+*
+* This source code is subject to the terms of the BSD 2 Clause License and
+* the Alliance for Open Media Patent License 1.0. If the BSD 2 Clause License
+* was not distributed with this source code in the LICENSE file, you can
+* obtain it at www.aomedia.org/license/software. If the Alliance for Open
+* Media Patent License 1.0 was not distributed with this source code in the
+* PATENTS file, you can obtain it at www.aomedia.org/license/patent.
+*/
+
+#ifndef EbGlobalMotionEstimationCost_h
+#define EbGlobalMotionEstimationCost_h
+
+#include "EbDefinitions.h"
+
+
+int gm_get_params_cost(const EbWarpedMotionParams *gm,
+                       const EbWarpedMotionParams *ref_gm, int allow_hp);
+
+#endif