This PR adds the support of global warped motion compensation for single and compound mode.
It is enabled only for 8bit streams with enc-mode 0.
The code is based on the libaom implementation with fastfeat.



Metric | deviation
-- | --
BD-PSNR | 0.008736805769833
BD-RATE | -0.185963014925564
Speed | -5.62%