@kylophone 2019-09-05
SVT-AV1 PoC global warped motion compensation experiments |   |   |   |   |   |   |   |   |   |   |   |  
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
Options: enc-mode 7 and variable q | CPU : Intel(R) Core(TM) i7-7700 CPU @ 3.60GH |   |   |   |   |   |   |   |   |   |   |  
  |   |   |   |   |   |   |   |   |   |   |   |  
  |   | q | size | PSNR Y | PSNR U | PSNR V | PSNR avg | PSNR min | PSNR max | Time | Time increase | Size evol
blue_sky_360p_60f.y4m | warped GMC #6d4717ef7f8ff1191713968d7624cd529b390f78 | 10 | 1113352 | 47.314821 | 47.722401 | 48.889715 | 47.609599 | 46.641002 | 51.87212 | 9572 | 507.26% | -0.20%
  |   | 20 | 576452 | 43.92826 | 45.145521 | 46.72919 | 44.483216 | 42.929579 | 48.528158 | 9543 | 564.34% | -0.73%
  |   | 30 | 333277 | 41.287576 | 43.403573 | 44.970259 | 42.042533 | 39.804811 | 45.457891 | 9465 | 617.01% | -0.75%
  |   | 40 | 180060 | 38.218857 | 40.970949 | 42.502124 | 39.09522 | 36.435064 | 42.43863 | 9206 | 669.53% | -1.40%
  |   | 50 | 101877 | 35.106358 | 38.292003 | 40.051139 | 36.075016 | 33.494281 | 39.28939 | 9032 | 731.93% | -1.64%
  |   | 60 | 57425 | 31.844284 | 35.756269 | 38.177377 | 32.961645 | 30.625647 | 35.916487 | 8918 | 762.87% | -1.97%
  |   |   |   |   |   |   |   |   |   |   |   |  
  | vanilla #f88378ef89dde81922de078b6342b8e0d948a7c0 | 10 | 1115615 | 47.310665 | 47.688081 | 48.867946 | 47.598343 | 46.624438 | 51.870128 | 1887 |   |  
  |   | 20 | 580700 | 43.933787 | 45.119354 | 46.703725 | 44.481108 | 42.960289 | 48.528158 | 1691 |   |  
  |   | 30 | 335800 | 41.27996 | 43.385793 | 44.975098 | 42.034734 | 39.729013 | 45.448022 | 1534 |   |  
  |   | 40 | 182615 | 38.216583 | 40.960998 | 42.491722 | 39.091497 | 36.19533 | 42.439606 | 1375 |   |  
  |   | 50 | 103572 | 35.088343 | 38.310286 | 40.048207 | 36.061624 | 33.246249 | 39.302144 | 1234 |   |  
  |   | 60 | 58580 | 31.807997 | 35.765452 | 38.172383 | 32.930887 | 30.339516 | 35.893766 | 1169 |   |  
  |   |   |   |   |   |   |   |   |   |   |   |  
shields_640x360_60f.y4m | warped GMC #6d4717ef7f8ff1191713968d7624cd529b390f78 | 10 | 1478779 | 44.131983 | 46.29982 | 47.724774 | 44.885504 | 43.744735 | 50.466383 | 6653 | 337.89% | 0.03%
  |   | 20 | 576936 | 41.319981 | 44.900881 | 46.455271 | 42.33916 | 40.889956 | 48.208105 | 6227 | 379.46% | 0.20%
  |   | 30 | 332331 | 39.636258 | 43.486881 | 44.947794 | 40.690904 | 38.719383 | 48.10015 | 6062 | 403.06% | 0.25%
  |   | 40 | 185370 | 37.250864 | 41.130033 | 42.619984 | 38.311583 | 36.464482 | 48.041969 | 6047 | 442.68% | 0.12%
  |   | 50 | 111817 | 34.701547 | 38.842693 | 40.403706 | 35.804426 | 34.150253 | 47.930199 | 5880 | 455.11% | 0.43%
  |   | 60 | 65611 | 31.748583 | 36.523239 | 37.963489 | 32.928909 | 31.234957 | 47.739624 | 5769 | 488.07% | 0.29%
  |   |   |   |   |   |   |   |   |   |   |   |  
  | vanilla #f88378ef89dde81922de078b6342b8e0d948a7c0 | 10 | 1478311 | 44.13242 | 46.301392 | 47.725818 | 44.88613 | 43.751514 | 50.437389 | 1969 |   |  
  |   | 20 | 575761 | 41.317483 | 44.901687 | 46.447433 | 42.336622 | 40.88125 | 48.208105 | 1641 |   |  
  |   | 30 | 331486 | 39.628675 | 43.495042 | 44.953328 | 40.685516 | 38.724227 | 48.10015 | 1504 |   |  
  |   | 40 | 185146 | 37.250551 | 41.125791 | 42.614044 | 38.310579 | 36.460471 | 48.041969 | 1366 |   |  
  |   | 50 | 111340 | 34.7053 | 38.849728 | 40.406704 | 35.808407 | 34.145398 | 47.930199 | 1292 |   |  
  |   | 60 | 65421 | 31.757373 | 36.516059 | 37.955197 | 32.935639 | 31.234957 | 47.739624 | 1182 |   |  
  |   |   |   |   |   |   |   |   |   |   |   |  
Netflix_PierSeaside_1920x1080_60fps_8bit_420_60f.y4m | warped GMC #6d4717ef7f8ff1191713968d7624cd529b390f78 | 10 | 5875075 | 46.940872 | 48.033376 | 46.466471 | 47.019011 | 46.233712 | 54.576071 | 32107 | 259.49% | -0.72%
  |   | 20 | 2163175 | 45.162395 | 47.098254 | 45.230922 | 45.442922 | 44.190356 | 51.489188 | 30083 | 283.00% | -0.67%
  |   | 30 | 1070852 | 43.97884 | 46.569804 | 44.781471 | 44.450959 | 42.413169 | 49.185666 | 28586 | 320.36% | -0.46%
  |   | 40 | 516458 | 42.377936 | 45.694601 | 44.195206 | 43.063345 | 40.7703 | 46.669576 | 27675 | 343.28% | -0.29%
  |   | 50 | 291872 | 40.728418 | 44.637859 | 43.461878 | 41.573184 | 39.446634 | 44.428286 | 27345 | 371.23% | -0.46%
  |   | 60 | 132531 | 38.099311 | 42.36835 | 41.67403 | 39.056433 | 37.427078 | 40.47662 | 26858 | 381.23% | -0.71%
  |   |   |   |   |   |   |   |   |   |   |   |  
  | vanilla #f88378ef89dde81922de078b6342b8e0d948a7c0 | 10 | 5917559 | 46.937195 | 48.012194 | 46.465446 | 47.013522 | 46.228978 | 54.576071 | 12373 |   |  
  |   | 20 | 2177795 | 45.157662 | 47.083218 | 45.228248 | 45.437375 | 44.193556 | 51.489188 | 10630 |   |  
  |   | 30 | 1075754 | 43.974091 | 46.560985 | 44.775148 | 44.445551 | 42.424775 | 49.185666 | 8923 |   |  
  |   | 40 | 517964 | 42.373676 | 45.685728 | 44.191254 | 43.058705 | 40.78639 | 46.669576 | 8062 |   |  
  |   | 50 | 293229 | 40.712913 | 44.616597 | 43.449252 | 41.557515 | 39.436388 | 44.428286 | 7366 |   |  
  |   | 60 | 133481 | 38.084081 | 42.354498 | 41.664564 | 39.041835 | 37.405181 | 40.47662 | 7045 |   |  
  |   |   |   |   |   |   |   |   |   |   |   |  
Netflix_FoodMarket_1920x1080_60fps_8bit_420_60f.y4m | warped GMC #6d4717ef7f8ff1191713968d7624cd529b390f78 | 10 | 16624271 | 44.402078 | 46.718585 | 47.047458 | 45.082046 | 42.85611 | 50.440963 | 89080 | 527.60% | -0.02%
  |   | 20 | 4728019 | 41.115686 | 44.910573 | 45.597281 | 42.108665 | 41.406504 | 46.349068 | 84688 | 609.66% | 0.02%
  |   | 30 | 1815712 | 39.889852 | 44.20589 | 45.068497 | 40.974803 | 39.994164 | 43.282872 | 82749 | 722.19% | -0.05%
  |   | 40 | 783382 | 38.67138 | 43.369894 | 44.347768 | 39.816278 | 38.375081 | 41.200681 | 81361 | 810.61% | -0.08%
  |   | 50 | 394818 | 36.976963 | 42.352324 | 43.382373 | 38.208201 | 36.433059 | 40.206511 | 80438 | 881.51% | 0.00%
  |   | 60 | 197774 | 34.837245 | 41.22381 | 42.207048 | 36.171382 | 34.184289 | 38.701119 | 80006 | 950.19% | -0.09%
  |   |   |   |   |   |   |   |   |   |   |   |  
  | vanilla #f88378ef89dde81922de078b6342b8e0d948a7c0 | 10 | 16628276 | 44.401848 | 46.718649 | 47.047522 | 45.081881 | 42.855783 | 50.442225 | 16884 |   |  
  |   | 20 | 4727307 | 41.115079 | 44.909116 | 45.59581 | 42.107919 | 41.41343 | 46.349656 | 13891 |   |  
  |   | 30 | 1816675 | 39.891284 | 44.205712 | 45.067572 | 40.975954 | 40.001047 | 43.283342 | 11458 |   |  
  |   | 40 | 784018 | 38.672727 | 43.368684 | 44.344147 | 39.817145 | 38.373636 | 41.200598 | 10037 |   |  
  |   | 50 | 394811 | 36.973411 | 42.353022 | 43.379898 | 38.204975 | 36.437154 | 40.206735 | 9125 |   |  
  |   | 60 | 197950 | 34.838598 | 41.223918 | 42.207906 | 36.17265 | 34.202833 | 38.701162 | 8420 |   |  