@kylophone 2019-09-05
@magwyz, Do you also have BD-rates for these clips?