@kylophone 2019-09-05
> This PR is a draft version to open questions:
> 
> * Do you have any specific methods to evaluate the integration of a new tool?

It should at least match, but ideally exceed the gains given by the same tool in libaom. The speed increase should also be in line with libaom, but ideally faster.  Have you measured this tool on/off using libaom?

> * Which samples are you using?

The `objective-1-fast` set on AWCY is a good set of clips to test with.

> * What should be the performance to have this new tool merged?

See above.

> * At what encoder mode / in which cases should it be enabled?

Good question.  @chkngit?