@magwyz 2019-09-07
@kylophone Thank you for your answers.
I'll complete the branch, run new experiments, calculate BD-rate and compare with libaom gains.