@chkngit 2019-09-09
for now we want to get this for at least M0 mode. but based on speed cost we might keep it for faster modes. I guess let's focus on getting the expected BDR gain  for the  objective-1-fast aom set (we expect ~-5% gain for blusky clip).