@magwyz 2019-09-13
Do you have any specific methods and tools to measure the BD-rate gain?