@EwoutH 2019-09-20
I measured the performance impact on a 48 vCPU Cascade Lake system for all encoder modes on a 1080p 8-bit and 2160p 10-bit clip, but the latter failed unfortunately.

| FPS (1080p 8b) | master | PR596 | Increase |
|-------------|-------:|------:|---------:|
| -enc-mode 8 | 58,305 | 1,875 |  3009,6% |
| -enc-mode 7 | 28,208 | 1,850 |  1424,8% |
| -enc-mode 6 | 21,116 | 1,775 |  1089,6% |
| -enc-mode 5 | 12,087 | 1,961 |   516,4% |
| -enc-mode 4 |  5,794 | 1,702 |   240,4% |
| -enc-mode 3 |  3,048 | 1,367 |   123,0% |
| -enc-mode 2 |  1,458 | 0,913 |    59,7% |
| -enc-mode 1 |  0,952 | 0,642 |    48,3% |
| -enc-mode 0 |  0,324 | 0,266 |    21,8% |

Since it's quite a big performance impact I would recommend only enabling this PR in m0, depending on bit-rate improvements.